requirements.txt
annotated-types==0.7.0
black==24.8.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==29.0.0
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
